requirements.txt
python-dotenv==1.0.1
pypdf==4.3.1
beautifulsoup4==4.12.3
requests==2.32.3
sentence-transformers==3.0.1
faiss-cpu==1.8.0
chromadb==0.5.3
transformers==4.44.2
huggingface-hub==0.24.6
accelerate==0.33.0
torch==2.2.0
tiktoken==0.7.0
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.38.0
httpx==0.27.2
langchain==0.2.14
langchain-community==0.2.12
numpy==1.24.0
pandas==2.0.0
scikit-learn==1.3.0
peft>=0.10.0
accelerate>=0.29.0
langchainhub>=0.1.14
langchain-chroma>=0.1.0
datasets>=2.14.0

